image_steganography/src/utils/config.py
"""
Configuration management utilities.

Handles loading and validating configuration files.
"""

import yaml
from pathlib import Path
from typing import Dict, Any


def load_config(config_path: str) -> Dict[str, Any]:
    """
    Load configuration from YAML file.
    
    Args:
        config_path (str): Path to config file
        
    Returns:
        Dict[str, Any]: Configuration dictionary
    """
    # TODO: Load and parse YAML config
    raise NotImplementedError


def validate_config(config: Dict[str, Any]) -> bool:
    """
    Validate configuration dictionary.
    
    Args:
        config (Dict[str, Any]): Configuration to validate
        
    Returns:
        bool: True if valid
    """
    # TODO: Validate required fields and values
    raise NotImplementedError


def save_config(config: Dict[str, Any], save_path: str):
    """
    Save configuration to YAML file.
    
    Args:
        config (Dict[str, Any]): Configuration to save
        save_path (str): Path to save config
    """
    # TODO: Save config to YAML
    raise NotImplementedError


class Config:
    """
    Configuration class with attribute access.
    """
    
    def __init__(self, config_dict: Dict[str, Any]):
        self._config = config_dict
        
    def __getattr__(self, name: str) -> Any:
        # TODO: Implement attribute access
        raise NotImplementedError
        
    def __setattr__(self, name: str, value: Any):
        # TODO: Implement attribute setting
        raise NotImplementedError
        
    def to_dict(self) -> Dict[str, Any]:
        """Convert to dictionary."""
        return self._config
