image_steganography/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0

# Training and monitoring
tensorboard>=2.13.0
tqdm>=4.65.0
wandb>=0.15.0

# Data handling
opencv-python>=4.7.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Utilities
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
